[FIXED JENKINS-25476 JENKINS-36375] Fix file corruption when updating files with Unicode BOM
This patch fixes the problem documented in
https://issues.jenkins-ci.org/browse/JENKINS-25476. I don't think this
is optimal solution, though. The Jenkins FilePath class should probably
provide a byte-oriented write method which could be used here.

But anyway I think Change Assembly Version plugin should adopt this
patch for now as it fixes the very real problem that is troubling
the plugin's typical use case where Jenkins is building projects
with AssemblyInfo.cs files that are occasionally updated with Visual
Studio, which in turn implies that by default BOMs are inserted in
the files.

Signed-off-by: Tuomas Jormola <tj@solitudo.net>